+++ src/Module/Api/Method/PlaylistSongsMethod.php
@@ -49,6 +49,7 @@ final class PlaylistSongsMethod
      * @param array $input
      * @param User $user
      * filter = (string) UID of playlist
+     * random = (integer) 0,1, if true get random songs using limit //optional
      * offset = (integer) //optional
      * limit  = (integer) //optional
      * @return boolean
@@ -61,6 +62,7 @@ final class PlaylistSongsMethod
         debug_event(self::class, 'User ' . $user->id . ' loading playlist: ' . $input['filter'], 5);
 
         $object_id = $input['filter'];
+        $random    = (array_key_exists('random', $input) && (int)$input['random'] == 1);
         $playlist  = ((int) $object_id === 0)
             ? new Search((int) str_replace('smart_', '', $object_id), 'song', $user)
             : new Playlist((int) $object_id);
@@ -77,7 +79,9 @@ final class PlaylistSongsMethod
             return false;
         }
 
-        $items = $playlist->get_items();
+        $items =($random)
+            ? $playlist->get_random_items()
+            : $playlist->get_items();
         if (empty($items)) {
             Api::empty('song', $input['api_format']);